I moved work regarding the save search button and filter button into another file
relates #112
